+++ domain/src/main/java/joe/barker/domain/boundary/MovieDetailsRepository.kt
@@ -5,5 +5,5 @@ import joe.barker.domain.entity.ErrorEntity
 import joe.barker.domain.entity.MovieDetails
 
 interface MovieDetailsRepository {
-    suspend fun getMovieDetailsOf(movieId: Int) : Either<MovieDetails?, ErrorEntity?>
+    suspend fun getMovieDetailsOf(movieId: Long) : Either<MovieDetails?, ErrorEntity?>
 }
